+++ index.js
@@ -3,19 +3,19 @@
 var Promise     = require('bluebird');
 var gitty       = require('gitty');
 var GithubAPI   = require('github');
-var tpAPI       = require('tp-api');
+var partial     = require('lodash.partial');
 var logger      = require('./lib/logging').logger();
 var Repository  = require('./lib/repository');
 var PullRequest = require('./lib/pull-request');
 var prMessage   = require('./lib/pull-request-message');
+var plugins     = require('./lib/plugins');
 
-var github, targetprocess;
+var github;
 
 module.exports = run;
 function run (config, options) {
   github = new GithubAPI({ version: '3.0.0' });
   github.authenticate(config.credentials.github);
-  targetprocess = tpAPI(config.credentials.targetprocess);
 
   var repo = new Repository(gitty);
 
@@ -25,29 +25,43 @@ function run (config, options) {
       repo.getCurrentBranchName(),
       repo.getRemoteOwnerName(),
       repo.getRemoteRepoName(),
-      prMessage(options.template, options.tp)
+      prMessage(options)
     ])
-    .spread( function (branch, user, repo, message) {
+    .spread( function (branch, user, repoName, message) {
+      options.pr = {
+        message : message
+      };
+
+      options.repo = {
+        branch : branch,
+        user   : user,
+        name   : repoName
+      };
+
       var pr   = new PullRequest(github, {
-        base     : options.base,
-        assignee : options.assignee,
-        branch   : branch,
-        user     : user,
-        repo     : repo
+        base     : options.cli.base,
+        assignee : options.cli.assignee,
+        branch   : options.repo.branch,
+        user     : options.repo.user,
+        repo     : options.repo.name
       });
 
-      createPullRequest(pr, message)
+      var context = {
+        pr: pr,
+        options: options
+      };
+
+      createPullRequest(context)
       .error(handle_cannotCreatePullRequest)
-      .then(assignPullRequest)
-      .then(commentOnTargetProcess(options.tp));
+      .then(partial(plugins.triggerPluginsForHook, 'afterCreatePullRequest', options));
     });
   });
 }
 
-function createPullRequest (pullRequest, message) {
-  return pullRequest.create({
-    title: message.title,
-    body: message.body
+function createPullRequest (context) {
+  return context.pr.create({
+    title : context.options.pr.message.title,
+    body  : context.options.pr.message.body
   });
 }
 
@@ -60,29 +74,4 @@ function handle_cannotCreatePullRequest (exception) {
   });
 
   process.exit(1);
-}
-
-function assignPullRequest (pullRequest) {
-  if (pullRequest.options.assignee) {
-    return pullRequest.assign();
-  } else {
-    return Promise.resolve(pullRequest);
-  }
-}
-
-function commentOnTargetProcess (tpEntityId) {
-  if (!tpEntityId) {
-    return function () {};
-  }
-
-  return function (pullRequest) {
-    var comment = 'Submitted Pull Request ' + pullRequest.url;
-    targetprocess().comment(tpEntityId, comment, function (error) {
-      if (error) {
-        logger.error(error);
-      } else {
-        logger.info('Commented on Target Process entity #', tpEntityId);
-      }
-    });
-  };
 }